Got first unittest in bootstrap/database/block.rn to pass.
Follow-on CLs will get the rest passing.  To get it to pass, we bind null(Foo) values to specific classes after initial binding, which is something we could not o with the old binder.  I also added not-null operators in code generators in several places.  Without them, it was creating multple versions of functions like destructors, one taking a non-nullable self, and another where self could be null.